refactor(即时列表): 函数重构
由油猴脚本改为CF workers反代获取数据

BREAKING CHANGE:
